last changes after Heorhii review
