+++ api/post.py
@@ -1,0 +1,13 @@
+from flask import request, jsonify, Blueprint
+
+from api.api_utils import get_payload
+
+post_api = Blueprint('post', __name__)
+
+
+@post_api.route('/post', methods=['POST'])
+def get_credentials():
+    token = request.headers['Authorization'].split()[1]
+    body = request.json
+    payload = get_payload(token)
+    return jsonify({'payload': payload, 'body': body})
