+++ config.py
@@ -2,4 +2,8 @@ import os
 
 from genersl_utils import read_key
 
-SECRET_KEY = os.environ.get('SECRET_KEY') or read_key('key.pub')
+
+class Config:
+
+    SECRET_KEY = os.environ.get('SECRET_KEY')
+    # SECRET_KEY = read_key('key.pub')
